+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/terminal/TerminalInterface.kt
@@ -5,6 +5,7 @@ import com.github.ajalt.mordant.input.MouseTracking
 import com.github.ajalt.mordant.rendering.AnsiLevel
 import com.github.ajalt.mordant.rendering.Size
 import kotlin.time.Duration
+import kotlin.time.TimeMark
 
 
 interface TerminalInterface {
@@ -48,8 +49,12 @@ interface TerminalInterface {
      *
      * You would return `null` if you receive an event that should be ignored, like a mouse
      * event when mouse tracking is off.
+     *
+     * @param timeout The point in time that this call should block to while waiting for an event.
+     *   If the timeout is in the past, this method should not block.
+     * @param mouseTracking The current mouse tracking mode.
      */
-    fun readInputEvent(timeout: Duration, mouseTracking: MouseTracking): InputEvent? {
+    fun readInputEvent(timeout: TimeMark, mouseTracking: MouseTracking): InputEvent? {
         throw NotImplementedError("Reading input events is not supported on this terminal")
     }